+++ spec/dummy/app/models/multi_applications_account.rb
@@ -1,3 +1,3 @@
 class MultiApplicationsAccount < ActiveRecord::Base
-  include BookingSync::Engine::Models::MultiApplicationsAccountModel
+  include BookingSync::Engine::Models::MultiApplicationsAccount
 end