+++ spec/dummy/db/migrate/20181130062531_create_multi_applications_accounts.rb
@@ -1,4 +1,4 @@
-class CreateMultiApplicationsAccounts < ActiveRecord::Migration[5.1]
+class CreateMultiApplicationsAccounts < ActiveRecord::Migration[5.2]
   def change
     create_table :multi_applications_accounts do |t|