+++ spec/dummy/db/migrate/20181130062650_add_o_auth_fields_to_multi_applications_accounts.rb
@@ -1,4 +1,4 @@
-class AddOAuthFieldsToMultiApplicationsAccounts < ActiveRecord::Migration[5.1]
+class AddOAuthFieldsToMultiApplicationsAccounts < ActiveRecord::Migration[5.2]
   def change
     add_column :multi_applications_accounts, :provider, :string
     add_column :multi_applications_accounts, :synced_id, :integer
@@ -8,6 +8,7 @@ class AddOAuthFieldsToMultiApplicationsAccounts < ActiveRecord::Migration[5.1]
     add_column :multi_applications_accounts, :oauth_refresh_token, :string
     add_column :multi_applications_accounts, :oauth_expires_at, :string
     add_column :multi_applications_accounts, :host, :string, null: false
-    add_index :multi_applications_accounts, :host, unique: true
+    add_index :multi_applications_accounts, :host
+    add_index :multi_applications_accounts, [:host, :synced_id], unique: true
   end
 end