+++ spec/models/multi_application_account_spec.rb
@@ -96,7 +96,7 @@ RSpec.describe MultiApplicationsAccount, type: :model do
   
     context "when the stored token is expired" do
       # comparing rails version, the use_transactional_fixtures only works pre 5
-      if Rails::VERSION::STRING.split(".").first.to_i >=5
+      if rails_version >=5
         self.use_transactional_tests = false
       else
         self.use_transactional_fixtures = false
@@ -215,6 +215,17 @@ RSpec.describe MultiApplicationsAccount, type: :model do
       expect(account.oauth_access_token).to eq "new_access_token"
       expect(account.oauth_refresh_token).to eq "new_refresh_token"
       expect(account.oauth_expires_at).to eq "new_expires_at"
+    end
+  end
+
+  describe ".find_by_host_and_synced_id" do
+    let!(:account_1) { MultiApplicationsAccount.create!(synced_id: 1, host: "example.test") }
+    let!(:account_2) { MultiApplicationsAccount.create!(synced_id: 2, host: "example.test") }
+    let!(:account_3) { MultiApplicationsAccount.create!(synced_id: 1, host: "example2.test") }
+    let!(:account_4) { MultiApplicationsAccount.create!(synced_id: 2, host: "example2.test") }
+
+    it "returns the right account" do
+      expect(MultiApplicationsAccount.find_by_host_and_synced_id("example2.test", 1)).to eq account_3
     end
   end
 end